+++ AtomEventStore.UnitTests/AtomEventStore.UnitTests.csproj
@@ -77,6 +77,7 @@
     <Compile Include="AtomAuthorTests.cs" />
     <Compile Include="AtomEntryBuilder.cs" />
     <Compile Include="AtomEntryComparer.cs" />
+    <Compile Include="AtomEntryLikeness.cs" />
     <Compile Include="AtomEntryTests.cs" />
     <Compile Include="AtomEnvy.cs" />
     <Compile Include="AtomEventObserverTests.cs" />
